+++ src/views/Profile/Profile.module.css
@@ -7,82 +7,8 @@
   overflow-y: auto;
   padding: var(--unit);
 }
-.contentWrap {
-  flex: 1 1 auto;
-  display: flex;
-  align-items: center;
-}
-.notFlexParent {
-  width: 100%;
-}
-
-.flexWrap {
-  flex: 1 0 auto;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin: auto 0;
-}
-
-.profileWrap {
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-  max-width: var(--section);
-  min-width: var(--min-section);
-}
-
-.divider {
-  flex: 1 1 auto;
-  max-width: 160px;
-}
-
-.settingsWrap {
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-  max-width: var(--section);
-  min-width: var(--min-section);
-}
 
 .loginWrap {
   margin: auto;
   max-width: var(--section);
-}
-
-.authInput:-webkit-autofill {
-  -webkit-background-clip: text;
-}
-
-.biography {
-  max-width: var(--section);
-}
-
-.saveWrap {
-  margin-top: var(--double-unit);
-  text-align: end;
-}
-
-.spinnerContainer {
-  /* same height as button it's replacing, cosmetic addition */
-  height: 42px;
-  & > svg {
-    height: 32px;
-  }
-}
-
-.logoutWrap {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  & > button {
-    margin: var(--unit);
-  }
-}
-
-/* Mobile view */
-@media screen and (max-width: 550px) {
-  .divider {
-    display: none;
-  }
 }
